BUG FIXED: pointer to stack-allocated data retrurned; BUG FIXED: size of browsers list computed incorrectly; STYLE: use size_t instead of int for array length

+++ MOTIF/BMotif_detect.h
@@ -29,7 +29,7 @@ OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED O
 #define H_DETECT
 
 extern void canonical(char*);
-extern char* detectApplications(char**,int);
+extern char* detectApplications(char**,size_t);
 extern char* detectEditors();
 extern char* detectBrowsers();
 extern char* getCommandExecutableName(char*);
